requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
backoff==2.2.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.5.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.4.0
chromadb==0.3.26
click==8.1.6
clickhouse-connect==0.8.9
# cloud-init
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.8
confluent-kafka==2.2.0
constantly==23.10.4
contourpy==1.3.0
cryptography
cycler==0.12.1
dataclasses-json==0.6.7
# dbus-python==1.3.2
deepdiff==8.0.1
# distro==1.9.0
# distro-info==1.7+build1
dnspython==2.7.0
docx==0.2.4
docx2txt==0.8
duckdb==1.1.3
effdet==0.4.1
elastic-transport==8.15.1
elasticsearch==8.16.0
emoji==2.14.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi>=0.100.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet==3.1.1
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.68.1
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.14.0
hiredis==3.0.0
hnswlib==0.8.0
httpcore==1.0.7
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib_resources==6.4.5
incremental==22.10.0
iopath==0.1.10
Jinja2==3.1.2
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.7
langchain
langchain-community
langchain-core
langchain-elasticsearch
langchain-experimental
langchain-google-genai
langchain-openai
langchain-text-splitters
langdetect==1.0.9
langsmith==0.1.147
launchpadlib==1.11.0
layoutparser==0.3.4
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx
onnxruntime
openai==1.56.1
opencv-python==4.9.0.80
openpyxl==3.1.2
orderly-set==5.2.2
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==10.2.0
pillow_heif==0.15.0
pinecone-client==2.2.1
portalocker==3.0.0
posthog==3.7.4
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
psycopg2==2.9.7
psycopg2-binary==2.9.10
pulsar-client==3.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.8
pycurl==7.45.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==5.1.0
PyPDF2
pypdfium2==4.30.0
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.7.7+ubuntu3
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.1
RapidFuzz==3.10.1
redis==4.6.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.5
scipy==1.13.1
selenium==4.24.0
service-identity==24.1.0
simsimd==6.2.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.13.1
systemd-python==235
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.7.0
timm==1.0.12
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
Twisted==24.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
unattended-upgrades==0.1
unstructured==0.15.9
unstructured-client==0.25.9
unstructured-inference==0.7.24
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.22.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
zope.interface==6.1
zstandard==0.23.0
